+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+- Fix some ST functions not being migrated to the GeometryType enum.
+
 ## [1.0.1](https://github.com/clickbar/laravel-magellan/tree/1.0.1) - 2023-01-05
 
 ### Fixed